rturn to float returned

+++ mymath/mymath.go
@@ -2,18 +2,18 @@ package mymath
 
 import "math"
 
-func Sqrt(x float64) int {
-	return int(math.Sqrt(x))
+func Sqrt(x float64) float64 {
+	return math.Sqrt(x)
 }
 
-func Abs(x float64) int {
-	return int(math.Abs(x))
+func Abs(x float64) float64 {
+	return math.Abs(x)
 }
 
-func Max(a float64, b float64) int {
-	return int(math.Max(a, b))
+func Max(a float64, b float64) float64 {
+	return math.Max(a, b)
 }
 
-func Yn(n int, x float64) int {
-	return int(math.Yn(n, x))
+func Yn(n int, x float64) float64 {
+	return math.Yn(n, x)
 }
